Add support for pyrogram v2
- Add support for generating new pyrogram string sessions in version 2.

-  Upgrade bot to pyrogram v2.

+++ StringSessionBot/database/users_sql.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, BigInteger
 from StringSessionBot.database import BASE, SESSION
 
 
 class Users(BASE):
     __tablename__ = "users"
     __table_args__ = {'extend_existing': True}
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
 
     def __init__(self, user_id, channels=None):
         self.user_id = user_id
